qcacld-3.0: Change pool mask to uint32_t from unsigned long
There is a crash observed while turning on the wifi.
In "hdd_apply_cfg_ini" api value from uint32_t type is
getting copied to unsigned long values, which is resulting
in a crash.

Change the type of ini values "provisioned_intf_pool" and
"derived_intf_pool" from unsigned long to uint32_t.

CRs-Fixed: 2159801

+++ core/hdd/inc/wlan_hdd_cfg.h
@@ -14766,8 +14766,8 @@ struct hdd_config {
 	uint8_t tx_chain_mask_5g;
 	uint8_t rx_chain_mask_5g;
 	bool mac_provision;
-	unsigned long provisioned_intf_pool;
-	unsigned long derived_intf_pool;
+	uint32_t provisioned_intf_pool;
+	uint32_t derived_intf_pool;
 };
 
 #define VAR_OFFSET(_Struct, _Var) (offsetof(_Struct, _Var))
